requirements.txt
Django==4.0.5
djangorestframework==3.13.1
python-decouple==3.6
celery==5.2.7
psycopg2-binary==2.9.3
selenium==4.2.0
django-filter==21.1
redis==4.3.3

